% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateData.R
\name{PoiBeta}
\alias{PoiBeta}
\title{PoiBeta}
\usage{
PoiBeta(k, a, b, n = 100, d = 0)
}
\arguments{
\item{k}{\code{numeric} vector for k.
if length > 1 then several rows are computed}

\item{a}{\code{numeric} vector for a}

\item{b}{\code{numeric} vector for b}

\item{n}{\code{integer} number of values (number of columns in value)}

\item{d}{\code{numeric} vector for probability of zero counts}
}
\value{
\code{numeric} matrix of length k, a, or b rows and n columns
}
\description{
Create Poisson-Beta distributions.
}
\details{
The Poisson-Beta distribution has mean and variance of
\itemize{
  \item Mu = k * a / (a + b)
  \item Sig^2 = k^2 * a * b / ((a + b + 1) * (a + b)^2)
}
If k, a, or b have length greater 1, a matrix with n columns is created.
}
