% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimFDRcontrol.R
\name{EstimFDRcontrol}
\alias{EstimFDRcontrol}
\title{EstimFDRcontrol}
\usage{
EstimFDRcontrol(dt, batches, groups, FDR)
}
\arguments{
\item{dt}{\code{num} matrix or data.frame of (not log transformed)
expression values with rows as genes and columns as cells}

\item{batches}{vector identifying batches (\code{chr}, or \code{int})}

\item{groups}{vector identifying batches (\code{chr}, or \code{int})}

\item{FDR}{\code{num} scalar of predicted FDR}
}
\value{
\code{num} for true (corrected) FDR
}
\description{
Estimate loss of FDR control for scDD using pooled cells and the 
Kolmogorov-Smirnov test.
Quality controlled, library sized normalized (and not log transformed)
values should be used.
}
\details{
When measurement comming from different batches are treated 
equally independent (pooling cells), then methods for detecting 
differentially distributed genes lose control over the false discovery rate
(FDR).
scDD is moderately robust to this effect, yet quite powerful.
This function uses your dataset to predict how much scDD underestimates
the FDR.
The predicted true FDR is returned (corrected for scDD's FDR control loss).

scDD is one of the most robust method with regard to batch effects
(when pooling cells).
If it's loss of FDR control is to high then you need to resort to drastic
methods such as proposed by A. Lun, Biostatistics, 2017.
Here, cells are summed up within batches and then edgeR is used with these
sums.
That means you effectively have to degrade your single cell experiment
to a bulk experiment.

If scDD's FDR control is within an acceptable range, we encourage using
the above mentioned method with edgeR (Lun, 2017) and scDD together.
scDD is quite good at detecting different shapes in distributions while
edgeR is good at detecting different means.
}
\note{
A linear mixed model is fitted iteratively to each gene, so this
function can take a while.
}
\examples{
ds <- SimulateData()
trueFDR <- EstimFDRcontrol(ds$table, ds$pData$batch, ds$pData$group, 0.1)
trueFDR

}
